requirements.txt
PyPDF2==3.0.1
Pillow==10.4.0
pypdfium2==4.30.0
